+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* [feature] 🌟 File multi-column anonymizer, inject sample rows in database from a CSV file.
+* [feature] 🌟 File enum anonymizer, inject samples in database from a plain text or CSV file.
 * [feature] 🌟 String pattern anonymizer, build complex strings by fetching values from other anonymizers.
 
 ## 2.0.3